+++ src/net/naonedbus/manager/impl/StopViewManager.java
@@ -142,7 +142,7 @@ public class StopViewManager extends SQLiteManager<Stop> {
 		mBuilder.setLongitude(c.getFloat(mColLongitude));
 		mBuilder.setIdStation(c.getInt(mColEquipmentId));
 		mBuilder.setOrdre(c.getInt(mColOrder));
-		mBuilder.setStepType(c.getInt(mColStepType));
+		mBuilder.setStepType(c.getLong(mColStepType));
 		return mBuilder.build();
 	}
 
@@ -158,7 +158,7 @@ public class StopViewManager extends SQLiteManager<Stop> {
 		mBuilder.setLongitude(c.getFloat(c.getColumnIndex(StopsViewTable.LONGITUDE)));
 		mBuilder.setIdStation(c.getInt(c.getColumnIndex(StopsViewTable.EQUIPMENT_ID)));
 		mBuilder.setOrdre(c.getInt(c.getColumnIndex(StopsViewTable.STOP_ORDER)));
-		mBuilder.setStepType(c.getInt(c.getColumnIndex(StopsViewTable.STEP_TYPE)));
+		mBuilder.setStepType(c.getLong(c.getColumnIndex(StopsViewTable.STEP_TYPE)));
 		return mBuilder.build();
 	}
